refactor: Service 단위 테스트 설정 클래스
